requirement.txt
openai
python-dotenv
PyMuPDF
tqdm
Flask
sentence-transformers
scikit-learn
numpy